refactor: split S3 value helpers
